+++ pkg/providers/papi/testdata/basic_with_production_activation/property.tf
@@ -47,18 +47,19 @@ resource "akamai_property" "test-edgesuite-net" {
   rules       = data.akamai_property_rules_template.rules.json
 }
 
-resource "akamai_property_activation" "test-edgesuite-net-staging" {
-  property_id                    = akamai_property.test-edgesuite-net.id
-  contact                        = ["jsmith@akamai.com"]
-  version                        = akamai_property.test-edgesuite-net.staging_version
-  network                        = "STAGING"
-  auto_acknowledge_rule_warnings = false
-}
-
-#resource "akamai_property_activation" "test-edgesuite-net-production" {
+#resource "akamai_property_activation" "test-edgesuite-net-staging" {
 #  property_id                    = akamai_property.test-edgesuite-net.id
 #  contact                        = []
 #  version                        = akamai_property.test-edgesuite-net.latest_version
-#  network                        = "PRODUCTION"
+#  network                        = "STAGING"
 #  auto_acknowledge_rule_warnings = false
 #}
+
+resource "akamai_property_activation" "test-edgesuite-net-production" {
+  property_id                    = akamai_property.test-edgesuite-net.id
+  contact                        = ["jsmith@akamai.com", "rjohnson@akamai.com"]
+  version                        = akamai_property.test-edgesuite-net.production_version
+  network                        = "PRODUCTION"
+  note                           = "example production note"
+  auto_acknowledge_rule_warnings = false
+}